pyproject.toml
[project]
name = "banter-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "elevenlabs>=2.36.1",
    "fastapi[standard]>=0.129.0",
    "openai>=2.21.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.1",
]
